Enable/disable mock mode from client
Before this commit, it was required to start the Mountebank server
with the `--mock` option to be able to use mock verification. As
explained [here](http://www.mbtest.org/docs/api/mocks), it is
possible to enable the mock mode on an imposter by creating
the imposter with the `recordRequest` option set to `true`.

Making this option available from the Mountebank .net client
makes it easier for the test writer to enable mock verification
without having to change the startup options of the mountebank
server that may already be running.